app/requirements.txt
streamlit
ultralytics
sentence-transformers
torch
torchvision
pillow
requests
transformers
chromadb